# BE-swagger.yaml
openapi: 3.0.4
info:
  title: Blood Donation Support System - OpenAPI 1.0
  version: v1
servers:
  - url: '/'
  - url: http://localhost:8080
  - url: https://be-t8i8.onrender.com
...
paths:
...
  /api/blood-units/{id}:
    get:
      tags:
        - Blood Units
      summary: Get Blood Units by Donation Process Id
      description: Get Blood Units by Donation Process Id
      operationId: getBloodUnitsByDonationProcessId
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: MongoId
      responses:
        default:
          description: Get blood units by donation process id success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BloodUnits'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
    patch:
      tags:
        - Blood Units
      summary: Update Blood Units by Donation Process Id for Staff or Admin
      description: Blood Units by Donation Process Id for Staff or Admin
      operationId: updateBloodUnitsByProcessId
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: MongoId
      requestBody:
        description: Information update health check
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/UpdateBloodUnitReqBody'
        required: true
      responses:
        default:
          description: Update health check success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BloodUnits'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
...
  /api/blood-units/{blood_group_id}/{blood_component_ids}:
    get:
      tags:
        - Blood Units
      summary: Get All Blood Units Relative
      description: Get blood units matching blood group and blood components
      operationId: getAllBloodUnitsRelative
      parameters:
        - name: blood_group_id
          in: path
          required: true
          schema:
            type: string
            format: MongoId

        - name: blood_component_ids
          in: path
          required: true
          style: simple # Comma-separated value
          explode: false # Important: keep as a single string
          schema:
            type: array
            items:
              type: string
              format: MongoId
      responses:
        '200':
          description: Blood units fetched successfully
          content:
            application/json:
              schema:
                type: array
                items:
                $ref: '#/components/schemas/BloodUnits'
        '401':
          description: Unauthorized
      security:
        - BearerAuth: []
  /api/donations/donation-registrations:
    post:
      tags:
        - Donation Registrations
      summary: Create Donation Registration
      description: Create a new donation registration
      operationId: createDonationRegistration
      requestBody:
        description: Information create donation registration
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDonationRegistrationReqBody'
        required: true
      responses:
        '200':
          description: Create donation registration success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationRegistration'
        '422':
          description: Invalid Input
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []

    get:
      tags:
        - Donation Registrations
      summary: Get All Donation Registrations for Staff or Admin
      description: Get all donation registrations for staff or admin
      operationId: getAllDonationRegistrations
      responses:
        default:
          description: Get all donation registrations success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DonationRegistration'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
  /api/donations/donation-registrations/user:
    get:
      tags:
        - Donation Registrations
      summary: Get Donation Registration by User ID Now
      description: Get donation registration by user id now
      operationId: getDonationRegistrationByUserId
      responses:
        default:
          description: Get donation registration by user id success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationRegistration'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
  /api/donations/donation-registrations/{id}:
    get:
      tags:
        - Donation Registrations
      summary: Get Donation Registration By Id
      description: Get Donation Registration By Id
      operationId: getDonationRegistrationById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: MongoId
      responses:
        default:
          description: Update donation registration success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationRegistration'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []

    patch:
      tags:
        - Donation Registrations
      summary: Update Donation Registration By Id
      description: Update donation registration By Id
      operationId: updateDonationRegistration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: MongoId
      requestBody:
        description: Information update donation registration
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDonationRegistrationReqBody'
        required: true
      responses:
        default:
          description: Update donation registration success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationRegistration'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
  /api/donations/donation-processes:
    get:
      tags:
        - Donation Processes
      summary: Get All Donation Processes for Staff or Admin
      description: Get all donation processes for staff or admin
      operationId: getAllDonationProcesses
      parameters:
        - in: query
          name: is_separated
          required: false
          schema:
            type: boolean
          description: Filter by separation status (true/false)
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [Pending, Approved, Rejected]
          description: Filter by donation process status
      responses:
        '200':
          description: Get all donation processes success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DonationProcess'
        '401':
          description: Unauthorized
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    DonationRegistrationStatus:
      type: string
      enum: [Pending, Approved, Rejected, Checked In]
      example: Pending
    DonationProcessStatus:
      type: string
      enum: [Pending, Approved, Rejected]
      example: Pending
    DonationType:
      type: string
      enum:
        - Whole Blood # Hiến máu toàn phần
        - Platelets # Hiến tiểu cầu
        - Plasma # Hiến huyết tương
        - Red Blood Cells # Hiến hồng cầu kép
        - Platelets - Plasma # Hiến tiểu cầu + huyết tương
        - Plasma - Red Blood Cells # Hiến huyết tương + hồng cầu
        - Platelets - Red Blood Cells # Hiến tiểu cầu + hồng cầu
      example: Whole Blood
...
    BloodUnitsStatus:
      type: string
      enum: [Available, Reserved, Used, Expired, Damaged]
      example: Expired
    BloodUnits:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        donation_process_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        request_process_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        blood_group_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        blood_component_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        volume:
          type: number
          example: 0
        donation_type:
          $ref: '#/components/schemas/DonationType'
        status:
          $ref: '#/components/schemas/BloodUnitsStatus'
        update_by:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        expired_at:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        created_at:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        updated_at:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
    UpdateBloodUnitReqBody:
      type: object
      properties:
        blood_group_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        blood_component_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        note:
          type: string
        volume:
          type: number
          example: 0
        storage_temperature:
          type: number
          example: 0
        status:
          $ref: '#/components/schemas/BloodUnitsStatus'
...
    CreateDonationRegistrationReqBody:
      type: object
      properties:
        blood_group_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        start_date_donation:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        donation_type:
          $ref: '#/components/schemas/DonationType'
        answers:
          type: array
          description: Danh sách câu trả lời khảo sát
          items:
            type: object
            properties:
              question_id:
                type: string
                format: MongoId
                example: 688947fb6fd8692c52479fe8
              answer:
                type: boolean
                example: true
    DonationRegistration:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        user_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        blood_group_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        blood_group_name:
          type: string
          example: 'B+'
        citizen_id_number:
          type: string
          example: 123456789123
        phone:
          type: string
          example: 0987654321
        full_name:
          type: string
          example: 'Thao Nguyen'
        blood_component_ids:
          type: array
          format: MongoId
          example: [682f6a3fdb3fb589d23e7654, 682f6a3fdb3fb589d23e7654]
        donation_type:
          $ref: '#/components/schemas/DonationType'
        start_date_donation:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        status:
          $ref: '#/components/schemas/DonationRegistrationStatus'
        created_at:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        updated_at:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
    UpdateDonationRegistrationReqBody:
      type: object
      properties:
        blood_group_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        donation_type:
          $ref: '#/components/schemas/DonationType'
        start_date_donation:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        status:
          $ref: '#/components/schemas/DonationRegistrationStatus'
    DonationProcess:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        donation_registration_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        user_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        health_check_id:
          type: string
          format: MongoId
          example: 6856be913ee8cc9717842b3f
        blood_group_id:
          type: string
          format: MongoId
          example: 6835ee86cb638a298130ff63
        blood_group_name:
          type: string
          example: 'B+'
        donation_type:
          $ref: '#/components/schemas/DonationType'
        username:
          type: string
          example: 'Thao Nguyen'
        donation_date:
          type: string
          format: date-time
          example: 2025-06-21T14:15:45.866Z
        volume_collected:
          type: number
          example: 0
        status:
          $ref: '#/components/schemas/DonationProcessStatus'
        is_separated:
          type: boolean
          example: false
        description:
          type: string
          example: ''
        created_at:
          type: string
          format: date-time
          example: 2025-06-21T14:15:45.866Z
        updated_at:
          type: string
          format: date-time
          example: 2025-06-21T14:15:45.866Z
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT